Move /app/ to /documents/
