+++ src/lib/utils/tests/search.test.ts
@@ -7,7 +7,7 @@ import { userDocs, tag, kv, searchUrl, highlight } from "../search";
 
 describe("search utilities", () => {
   test("search URL", () => {
-    const href = new URL(`/app/`, APP_URL);
+    const href = new URL(`/documents/`, APP_URL);
     href.searchParams.set("q", "Nick Clegg");
     expect(searchUrl("Nick Clegg")).toStrictEqual(href);
   });
