--- src/routes/app/+layout.svelte
@@ -1,5 +1,0 @@
-<svelte:head>
-  <title>DocumentCloud</title>
-</svelte:head>
-
-<slot />
