+++ src/routes/documents/sidebar/AddOns.svelte
