+++ src/routes/upload/+layout.svelte
@@ -4,8 +4,8 @@
   import MainLayout from "@/lib/components/layouts/MainLayout.svelte";
   import SignedIn from "@/lib/components/common/SignedIn.svelte";
 
-  import Documents from "../app/sidebar/Documents.svelte";
-  import Projects from "../app/sidebar/Projects.svelte";
+  import Documents from "../documents/sidebar/Documents.svelte";
+  import Projects from "../documents/sidebar/Projects.svelte";
 </script>
 
 <MainLayout>
